Refactor time effects
This commit refactors the 'weather' and 'random tick speed' time effects
into extensions of a base class that TimeService can store in a list and
call in a generic way. This will allow myself and other mod creators to
more easily add additional sleep/time effects, and sets the mod up for a
time effect registry in the future.